+++ package.json
@@ -120,10 +120,10 @@
     "clean": "rimraf dist",
     "build:types": "tsc -p tsconfig.build.json --emitDeclarationOnly",
     "build:js": "tsup --config tsup.config.ts",
-    "build:abis": "ncp src/internal/abis dist/internal/abis",
+    "build:abis": "mkdir -p dist/core/internal/abis && cp -R src/core/internal/abis/* dist/core/internal/abis/",
     "build": "bun run clean && bun run build:types && bun run build:js && bun run build:abis",
     "typecheck": "tsc -p tsconfig.build.json --noEmit",
-    "types": "typechain --target ethers-v6 --out-dir src/adapters/ethers/typechain src/internal/abis/json/*.json && ncp src/adapters/ethers/typechain typechain",
+    "types": "typechain --target ethers-v6 --out-dir src/adapters/ethers/typechain src/core/internal/abis/*.json && ncp src/adapters/ethers/typechain typechain",
     "prepublishOnly": "bun run clean && bun run build",
     "lint": "eslint . --ext .ts,.tsx",
     "lint:fix": "eslint . --ext .ts,.tsx --fix",
