+++ src/adapters/ethers/__tests__/client.test.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect } from 'bun:test';
 import { Interface } from 'ethers';
 import { createEthersClient, type EthersClient } from '../client';
-import IBridgehubABI from '../../../internal/abis/IBridgehub.json';
-import IL1AssetRouterABI from '../../../internal/abis/IL1AssetRouter.json';
-import IL1NullifierABI from '../../../internal/abis/IL1Nullifier.json';
+import IBridgehubABI from '../../../core/internal/abis/IBridgehub.json';
+import IL1AssetRouterABI from '../../../core/internal/abis/IL1AssetRouter.json';
+import IL1NullifierABI from '../../../core/internal/abis/IL1Nullifier.json';
 import {
   L2_ASSET_ROUTER_ADDR,
   L2_NATIVE_TOKEN_VAULT_ADDR,
