+++ src/adapters/ethers/resources/deposits/routes/erc20-nonbase.ts
@@ -4,8 +4,8 @@ import type { DepositRouteStrategy } from './types';
 import { Contract } from 'ethers';
 import type { TransactionRequest } from 'ethers';
 import { encodeSecondBridgeErc20Args } from '../../utils';
-import IERC20ABI from '../../../../../internal/abis/IERC20.json' assert { type: 'json' };
-import IBridgehubABI from '../../../../../internal/abis/IBridgehub.json' assert { type: 'json' };
+import IERC20ABI from '../../../../../core/internal/abis/IERC20.json' assert { type: 'json' };
+import IBridgehubABI from '../../../../../core/internal/abis/IBridgehub.json' assert { type: 'json' };
 import type { ApprovalNeed, PlanStep } from '../../../../../core/types/flows/base';
 import { createErrorHandlers } from '../../../errors/error-ops';
 import { OP_DEPOSITS } from '../../../../../core/types';
