+++ src/adapters/ethers/resources/withdrawals/routes/eth.ts
@@ -3,7 +3,7 @@ import { Contract, Interface, type TransactionRequest } from 'ethers';
 import type { WithdrawRouteStrategy } from './types';
 import type { PlanStep } from '../../../../../core/types/flows/base';
 import { L2_BASE_TOKEN_SYSTEM_CONTRACT_ADDR } from '../../../../../core/constants';
-import L2BaseTokenABI from '../../../../../internal/abis/IBaseToken.json' assert { type: 'json' };
+import L2BaseTokenABI from '../../../../../core/internal/abis/IBaseToken.json' assert { type: 'json' };
 
 import { createErrorHandlers } from '../../../errors/error-ops';
 import { OP_WITHDRAWALS } from '../../../../../core/types';
