+++ src/adapters/viem/errors/revert.ts
@@ -2,11 +2,11 @@
 import type { Abi } from 'viem';
 import { decodeErrorResult } from 'viem';
 
-import IL1NullifierABI from '../../../internal/abis/IL1Nullifier.json' assert { type: 'json' };
-import IL1NativeTokenVaultABI from '../../../internal/abis/L1NativeTokenVault.json' assert { type: 'json' };
-import IL2NativeTokenVaultABI from '../../../internal/abis/L2NativeTokenVault.json' assert { type: 'json' };
-import IERC20ABI from '../../../internal/abis/IERC20.json' assert { type: 'json' };
-import MailboxABI from '../../../internal/abis/Mailbox.json' assert { type: 'json' };
+import IL1NullifierABI from '../../../core/internal/abis/IL1Nullifier.json' assert { type: 'json' };
+import IL1NativeTokenVaultABI from '../../../core/internal/abis/L1NativeTokenVault.json' assert { type: 'json' };
+import IL2NativeTokenVaultABI from '../../../core/internal/abis/L2NativeTokenVault.json' assert { type: 'json' };
+import IERC20ABI from '../../../core/internal/abis/IERC20.json' assert { type: 'json' };
+import MailboxABI from '../../../core/internal/abis/Mailbox.json' assert { type: 'json' };
 
 import { REVERT_TO_READINESS } from '../../../core/errors/withdrawal-revert-map';
 import type { FinalizeReadiness } from '../../../core/types/flows/withdrawals';
