+++ src/adapters/viem/sdk.ts
@@ -21,13 +21,13 @@ import {
 } from '../../core/constants';
 
 // ABIs (to type contract handles returned from helpers.contracts())
-import IBridgehubABI from '../../internal/abis/IBridgehub.json' assert { type: 'json' };
-import IL1AssetRouterABI from '../../internal/abis/IL1AssetRouter.json' assert { type: 'json' };
-import IL1NullifierABI from '../../internal/abis/IL1Nullifier.json' assert { type: 'json' };
-import L1NativeTokenVaultABI from '../../internal/abis/L1NativeTokenVault.json' assert { type: 'json' };
-import IL2AssetRouterABI from '../../internal/abis/IL2AssetRouter.json' assert { type: 'json' };
-import L2NativeTokenVaultABI from '../../internal/abis/L2NativeTokenVault.json' assert { type: 'json' };
-import IBaseTokenABI from '../../internal/abis/IBaseToken.json' assert { type: 'json' };
+import IBridgehubABI from '../../core/internal/abis/IBridgehub.json' assert { type: 'json' };
+import IL1AssetRouterABI from '../../core/internal/abis/IL1AssetRouter.json' assert { type: 'json' };
+import IL1NullifierABI from '../../core/internal/abis/IL1Nullifier.json' assert { type: 'json' };
+import L1NativeTokenVaultABI from '../../core/internal/abis/L1NativeTokenVault.json' assert { type: 'json' };
+import IL2AssetRouterABI from '../../core/internal/abis/IL2AssetRouter.json' assert { type: 'json' };
+import L2NativeTokenVaultABI from '../../core/internal/abis/L2NativeTokenVault.json' assert { type: 'json' };
+import IBaseTokenABI from '../../core/internal/abis/IBaseToken.json' assert { type: 'json' };
 
 // Helpers to express the contracts() return type
 type ViemContracts = {
