+++ src/core/internal/abis/IL1AssetRouter.json
