+++ src/core/internal/abis/L1NativeTokenVault.json
