+++ src/core/internal/abis/L2NativeTokenVault.json
